Patch by Craig Ringer: ======================
The attached patch alters PoDoFo's pdf name escaping
behaviour to ensure that two names with the same
meaning always compare equal, and to make it easier to
work with names.
    
I wrote the bulk of this patch before the existing PDF
name escaping support was committed to svn. I'm
submitting this because I think that support is
somewhat incomplete - in particular, two PdfName
objects with the same meaning but escaped two different
ways do not compare equal. Fixing this would require:
 - Unescaping both objects for each comparison
operation; or
 - Unescaping every name object on creation then
   re-escaping it in a standardised way
	       
... and since I already had this approach basically
done I thought I'd submit it instead.
		  
PdfName is changed to use an unescaped internal
representation. When constructed it treats the input
string as unescaped unless the
PdfName::FromEscaped(...) calls are used to construct a
PdfName object.
		        
GetName() returns the unescaped name; a new
GetEscapedName() is added to return the escaped name
for those uses who may need it. Write() is updated to
escape its output. Caching of the escaped
representation might be desirable if a memory/cpu
trade-off proves worthwhile, but this is not currently
done.
			        
In addition, I've switched PdfParserBase over to use
the IsWhiteSpace(...) and IsDelimeter(...) tests I used
in my older work. That change can easily be discarded
if you don't like it, but I found it a _lot_ faster.

Comment by domseichter:
=======================

Thanks for your patch. I commited it to SVN.

I am still thinking about what is the best solution for 
the PdfName problem. You are definitly right, that the old 
way of storing the escaped version might cause problems. 
Therefore I commit your patch.

The problem is though that storing the unescaped version 
is slower than storing the escaped version. On my system 
parsing the PDFReference16.pdf PDF file takes now an 
average of 26 seconds instead of 5 seconds before. This 
increase in parsing time is almost inacceptable. But an 
incorrectly working version is not better either.

So I am trying to find a better and correctly working 
solution for this (most likely unescaping before 
comparison). I have to check wether this is faster or not.

best regards,

 Dom





git-svn-id: https://podofo.svn.sourceforge.net/svnroot/podofo/podofo/trunk@231 f068efa2-cb12-0410-9b19-e61e8c05d287