- Make some more small and frequently used methods inlinable - Add a --enable-debug= argument to configure - Add a LOT more warning flags to the build command line - Fix a LOT of the resulting warnings    - Most importantly, (cstyle) casts now generate a warning
Note that this patch still dramatically increases the number of warnings generated
by a build, but not to the point where anything else is drowned like it did before
the builk of them were fixed. More fixes for these warnings will come in over time.
Warnings in the tests have not yet been cleaned up.



git-svn-id: https://podofo.svn.sourceforge.net/svnroot/podofo/podofo/trunk@241 f068efa2-cb12-0410-9b19-e61e8c05d287