Added a tokenizer for PDF file PdfTokenizer; Replaced PdfParserBase with PdfTokenizer; Simplified PdfVariant, PdfParser and PdfParserObject as all datatypes are now read by PdfTokenizer::GetNextVariant() and comments are handled under the way;
PdfTokenizer passes all tests, but has still problems with some encrypted PDF files and I currently have a problem that PdfWriter creates invalid XRef tables. I am not yet sure wether this is related to the new PdfTokenizer or not;

I am still commiting now, knowing of the above two problems, because PdfTokenizer makes the code a lot cleaner and is generally the better approach. We have now _ONE_ place where optimizations and bug fixes regarding reading of PdfObjects or datatypes can go.

Additionally PdfTokenizer will also work for parsing PDF Content streams;

I did not choose to use lex or flex for the lexing, as all the code was basically there already. I just had to create a (hopefully) nice API for PdfTokenizer and port all the code and tests to the new classes.




git-svn-id: https://podofo.svn.sourceforge.net/svnroot/podofo/podofo/trunk@319 f068efa2-cb12-0410-9b19-e61e8c05d287