PdfStreamedDocument works now, replacing the WriteImmediate stuff in PdfDocument. PdfImage's and annotations won't work yet with PdfDocument. The filters have precedence now though.
Additionally I fixed a crash in PdfOutlines and PdfDocument::Append merges now the outline structure of both documents correctly. 
I was able to append the PDFReference16.pdf to Apple's ObjectiveCbook to obtain a 1500 pages PDF file. The merge was successfull, 
only named destinations do not work in the merged PDF as nametrees are not yet merged.

To make PdfDocument::Append working operator= had to be implemented for PdfStream.



git-svn-id: https://podofo.svn.sourceforge.net/svnroot/podofo/podofo/trunk@411 f068efa2-cb12-0410-9b19-e61e8c05d287