FIXED: If PDF Stream writer writes an object (because stream data is appended), no key may be added to this object as all keys are already written and changes won't make it to the final PDF. Therefore certain order of calls to PoDoFo methods resulted in broken PDFs.
PdfVariants are now set to immutable once they are written by PdfImmediateWriter and the PdfDataTypes throw an exception if they are changed after 
being set to immutable. There might be places in our current code where such problems might occur. We will have to fix them as they are detected. 

This caused the problems reported by Jörg Richter on the mailing list.


git-svn-id: https://podofo.svn.sourceforge.net/svnroot/podofo/podofo/trunk@1160 f068efa2-cb12-0410-9b19-e61e8c05d287